+++ contracts/utility/MultiIncrementer.sol
@@ -1,6 +1,5 @@
 // SPDX-License-Identifier: MIT
 // Based on OpenZeppelin Contracts v4.4.1 (utils/Counters.sol)
-
 pragma solidity ^0.8.11;
 
 /**